fix(#617): remove better-sqlite3, convert tests to real PostgreSQL
Replace in-memory SQLite test doubles with real PostgreSQL testing.
Fix production bugs discovered during migration: placeholder conversion
for empty string literals, transaction pool routing, PG bigint/numeric
type casts, and identifier quoting. Add Zod compiler setup to dashboard
tests that was causing CI schema validation failures.

- Add test-db-helper.ts with getTestDb/truncateTestTables/closeTestDb
- Add init-test-db.sh for test database creation in docker-compose
- Convert traces-ingest, traces, and prometheus test files
- Fix convertPlaceholders() for '' empty SQL string literals
- Fix PostgresAdapter.transaction() to use injected pool
- Add ::integer/::float casts for PG COUNT/AVG return types
- Quote camelCase aliases in SQL for PG identifier lowercasing
- Add postgres-test service to CI workflow
- Remove better-sqlite3 from devDependencies

+++ backend/src/routes/prometheus.ts
@@ -96,7 +96,7 @@ async function getCachedSnapshot(): Promise<MetricsSnapshot> {
   const monitoringDb = getDbForDomain('monitoring');
 
   const insights = await insightsDb.query<{ severity: string; category: string; total: number }>(
-    `SELECT severity, category, COUNT(*) as total
+    `SELECT severity, category, COUNT(*)::integer as total
      FROM insights
      GROUP BY severity, category`,
     [],
@@ -110,7 +110,7 @@ async function getCachedSnapshot(): Promise<MetricsSnapshot> {
          WHEN lower(title) LIKE '%memory%' THEN 'memory'
          ELSE 'unknown'
        END as metric_type,
-       COUNT(*) as total
+       COUNT(*)::integer as total
      FROM insights
      WHERE category = 'anomaly'
        AND container_name IS NOT NULL
@@ -119,7 +119,7 @@ async function getCachedSnapshot(): Promise<MetricsSnapshot> {
   );
 
   const actions = await actionsDb.query<{ status: string; total: number }>(
-    `SELECT status, COUNT(*) as total
+    `SELECT status, COUNT(*)::integer as total
      FROM actions
      GROUP BY status`,
     [],
@@ -134,7 +134,7 @@ async function getCachedSnapshot(): Promise<MetricsSnapshot> {
   );
 
   const activeAnomalies = await insightsDb.queryOne<{ count: number }>(
-    `SELECT COUNT(*) as count
+    `SELECT COUNT(*)::integer as count
      FROM insights
      WHERE category = 'anomaly' AND is_acknowledged = false`,
     [],
